requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-extensions==3.2.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
channels==4.1.0
dj-rest-auth==6.0.0
django-rosetta==0.10.0
django-allauth==0.63.3
